requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
defusedxml==0.7.1
et-xmlfile==1.1.0
idna==3.4
jira==3.5.2
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
